trendsurf-copilot/requirements.txt
# TrendSurf Copilot — Dependencies
openai>=1.30.0
azure-identity>=1.21.0
python-dotenv>=1.1.0
